pyproject.toml
[tool.poetry]
name = "tasks-management-app"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.1"
slick-reporting = "^0.1.3"
django-ninja = "^1.3.0"
pydantic = {extras = ["email"], version = "^2.9.1"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
